SC1.2.2-video-captions (#158)
* Added Video captions rule and necessary test assets.

Added Video captions rule (to test if correct captions are available for video element) and necessary test assets (a cropped version of one of the w3c perspective video's).

* Fixed some file paths

Fixed some file paths

* video-captions: Editorial update

* Update video-captions

* Added 1 test case and descriptions for test cases

Added 1 test case (associated text with ariadescribedby under "failed") and descriptions for all the test cases.

* Added assumption

Added assumption that the video can be started by the user

* Moved 1.2.2 Media laternative for text into this PR

We started writing 1.2.2 Media alternative for text into the 1.2.1 PR because we also needed it for audio only. But we eventually chose to split it into three atomic rules, so now we moved it to the correct PR.

* Added composate rule for video alternatives to media

We added a composate rule for video alternatives to media. This combines the outcome of two rules (two methods of succeeding the SC)

* Deleted media alternative rule

Deleted the media alternative rule

* Changed some wordings and fixed some templating

Changed some wordings and fixed the rules to better adhere to the templates.

* Editorials for video caption rules