Reverting changes to "included in the acc. tree"
Removing the "expanded" description of when DOM nodes are included in the accessibility tree, since it was too simplified and only applied to native HTML elements, and not even covering all the edge cases for that technology. 
The source of truth for this lies within the specs already linked to from the definition.