scrollable-element-keyboard: inapplicable example 5 started failing (#1743)
* scrollable-element-keyboard: inapplicable example 5 started failing

Chrome 96 changed how it padding-right interacts with scrollbar. In previous versions, and still in all other major browser it would pretend right padding does not exist for the purpose of scrolling. Effectively just making the box wider by the amount of padding-right that it has, and pretending that padding-right is 0. That's no longer the case, and so this inapplicable example was suddenly a failure.

I've updated this example so that it functions consistently again in all major browsers.

* Update _rules/scrollable-element-keyboard-accessible-0ssw9k.md

Co-authored-by: Jean-Yves Moyen <jym@siteimprove.com>

* Update scrollable-element-keyboard-accessible-0ssw9k.md

Co-authored-by: Jean-Yves Moyen <jym@siteimprove.com>